refactor separate automation catalog from connector runtime
Split automation descriptors and Discord rule catalogs out of connector runtime code, keep the local Discord connector path, and remove the dead gateway wire layer so the bot runtime reflects the current architecture.
