--- flexus_client_kit/gateway/__init__.py
@@ -1,27 +1,0 @@
-from flexus_client_kit.gateway.ckit_gateway_wire import (
-    WIRE_V,
-    GatewayActionCommandEnvelope,
-    GatewayActionResultEnvelope,
-    action_result_from_dict,
-    action_result_to_dict,
-    channel_cmd_discord,
-    channel_reply_discord,
-    gateway_instance_key_from_token,
-    gateway_result_envelope_from_dict,
-    normalized_event_from_dict,
-    parse_action_command_envelope,
-)
-
-__all__ = [
-    "WIRE_V",
-    "GatewayActionCommandEnvelope",
-    "GatewayActionResultEnvelope",
-    "action_result_from_dict",
-    "action_result_to_dict",
-    "channel_cmd_discord",
-    "channel_reply_discord",
-    "gateway_instance_key_from_token",
-    "gateway_result_envelope_from_dict",
-    "normalized_event_from_dict",
-    "parse_action_command_envelope",
-]
